Move ViewTracker interface to base view tracker package
