+++ app/src/main/java/com/github/vase4kin/teamcityapp/account/create/tracker/CreateAccountTracker.java
@@ -16,7 +16,7 @@
 
 package com.github.vase4kin.teamcityapp.account.create.tracker;
 
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 
 /**
  * Tracking
