+++ app/src/main/java/com/github/vase4kin/teamcityapp/agents/dagger/AgentModule.java
@@ -31,7 +31,7 @@ import com.github.vase4kin.teamcityapp.agents.view.AgentsAdapter;
 import com.github.vase4kin.teamcityapp.api.Repository;
 import com.github.vase4kin.teamcityapp.base.list.view.BaseListView;
 import com.github.vase4kin.teamcityapp.base.list.view.ViewHolderFactory;
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 
 import java.util.Map;
