+++ app/src/main/java/com/github/vase4kin/teamcityapp/base/tracker/BaseViewTracker.java
@@ -16,8 +16,6 @@
 
 package com.github.vase4kin.teamcityapp.base.tracker;
 
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
-
 import java.util.Set;
 
 /**
