+++ app/src/main/java/com/github/vase4kin/teamcityapp/changes/presenter/ChangesPresenterImpl.java
@@ -21,13 +21,13 @@ import android.support.annotation.VisibleForTesting;
 
 import com.github.vase4kin.teamcityapp.account.create.data.OnLoadingListener;
 import com.github.vase4kin.teamcityapp.base.list.presenter.BaseListPresenterImpl;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 import com.github.vase4kin.teamcityapp.changes.api.Changes;
 import com.github.vase4kin.teamcityapp.changes.data.ChangesDataManager;
 import com.github.vase4kin.teamcityapp.changes.data.ChangesDataModel;
 import com.github.vase4kin.teamcityapp.changes.data.ChangesDataModelImpl;
 import com.github.vase4kin.teamcityapp.changes.extractor.ChangesValueExtractor;
 import com.github.vase4kin.teamcityapp.changes.view.ChangesView;
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
 import com.mugen.MugenCallbacks;
 
 import java.util.List;
