+++ app/src/main/java/com/github/vase4kin/teamcityapp/filter_builds/tracker/FilterBuildsTracker.java
@@ -17,7 +17,7 @@
 package com.github.vase4kin.teamcityapp.filter_builds.tracker;
 
 
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 
 /**
  * Filter builds tracking class
