+++ app/src/main/java/com/github/vase4kin/teamcityapp/runbuild/tracker/RunBuildTracker.java
@@ -16,7 +16,7 @@
 
 package com.github.vase4kin.teamcityapp.runbuild.tracker;
 
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 
 /**
  * Run build tracking
