+++ app/src/main/java/com/github/vase4kin/teamcityapp/tests/dagger/TestsModule.java
@@ -25,7 +25,7 @@ import com.github.vase4kin.teamcityapp.api.Repository;
 import com.github.vase4kin.teamcityapp.base.list.view.BaseListView;
 import com.github.vase4kin.teamcityapp.base.list.view.SimpleSectionedRecyclerViewAdapter;
 import com.github.vase4kin.teamcityapp.base.list.view.ViewHolderFactory;
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 import com.github.vase4kin.teamcityapp.tests.data.TestsDataManager;
 import com.github.vase4kin.teamcityapp.tests.data.TestsDataManagerImpl;
 import com.github.vase4kin.teamcityapp.tests.data.TestsDataModel;
