+++ app/src/main/java/com/github/vase4kin/teamcityapp/tests/presenter/TestsPresenterImpl.java
@@ -24,7 +24,7 @@ import android.view.MenuItem;
 
 import com.github.vase4kin.teamcityapp.account.create.data.OnLoadingListener;
 import com.github.vase4kin.teamcityapp.base.list.presenter.BaseListPresenterImpl;
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 import com.github.vase4kin.teamcityapp.tests.api.TestOccurrences;
 import com.github.vase4kin.teamcityapp.tests.data.TestsDataManager;
 import com.github.vase4kin.teamcityapp.tests.data.TestsDataModel;
