+++ app/src/test/java/com/github/vase4kin/teamcityapp/base/tabs/presenter/BaseTabsPresenterImplTest.java
@@ -18,7 +18,7 @@ package com.github.vase4kin.teamcityapp.base.tabs.presenter;
 
 import com.github.vase4kin.teamcityapp.base.tabs.data.BaseTabsDataManager;
 import com.github.vase4kin.teamcityapp.base.tabs.view.BaseTabsViewModel;
-import com.github.vase4kin.teamcityapp.navigation.tracker.ViewTracker;
+import com.github.vase4kin.teamcityapp.base.tracker.ViewTracker;
 
 import org.junit.Before;
 import org.junit.Test;
